moved lib folder under app
