+++ app/lib/graphql/queries/sc-labs/home.graphql
