+++ tests/test_sch_content.py
@@ -1,12 +1,12 @@
 import pytest
-from lib.content.sch_content import get_home
+from app.lib.content.sch_content import get_home
 
 def test_get_home(mocker):
     mocker.patch(
-        'lib.graphql.client.query'
+        'app.lib.graphql.client.query'
     )
     mocker.patch(
-        'lib.mapper.sch_mapper.map_home',
+        'app.lib.mapper.sch_mapper.map_home',
         return_value={'title': 'Search for benefits and services'}
     )
     home_content = get_home()
