+++ tests/test_sch_mapper.py
@@ -1,5 +1,5 @@
 import pytest
-from lib.mapper.sch_mapper import map_home
+from app.lib.mapper.sch_mapper import map_home
 
 def test_map_home():
     aem_content = {'alphaDCHomeByPath': {'item': {'_path': '/content/dam/decd-endc/content-fragments/alpha/dev/components/home/search-for-benefits-and-services', 'scTitleEn': 'Search for benefits and services'}}}
